requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
stripe==7.5.0
WTForms==3.1.0
email-validator==2.1.0